Updated module augmentation for heading package.
